Renamed redeemers
To be defined from the script's point of view.

- Current -> Unused
- Outdated -> Used
